backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
python-multipart
pydantic
scikit-learn
numpy
PyMuPDF
